+++ slices/sorted_index_by_test.go
@@ -85,6 +85,18 @@ func TestSortedIndexBy(t *testing.T) {
 			value: 3,
 			want:  3,
 		},
+		{
+			name:  "more more sequence",
+			arr:   []int{0, 1, 2, 3, 3, 3, 3, 3, 3, 4, 5, 6, 7, 8, 9},
+			value: 3,
+			want:  3,
+		},
+		{
+			name:  "at the end",
+			arr:   []int{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
+			value: 10,
+			want:  10,
+		},
 	}
 
 	for _, subject := range tests {